+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 7.0.0
+- Feature: Added CPU usage monitoring via `getCpuUsage`.
+- Feature: Added RAM usage monitoring via `memoryUsage`.
+
 ## 6.9.6
 - Remove: `getIosTunnelStatus()` expose.
 - Improvement[iOS]: `start()` now checks if the tunnel is already running and acts as